requirements.txt
aiohttp
aiosignal
anyio
asgiref
attrs
backoff
bcrypt
beautifulsoup4
build
cachetools
certifi
charset-normalizer
chromadb
click
colorama
coloredlogs
dataclasses-json
decorator
Deprecated
distro
duckduckgo_search
durationpy
exceptiongroup
#faiss-cpu
fastapi
filelock
filetype
fonttools
fsspec
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
google_search_results
googleapis-common-protos
grpcio
grpcio-status
h11
httpcore
httplib2
httptools
httpx
httpx-sse
huggingface-hub
idna
importlib_metadata
importlib_resources
ipykernel
ipython
jedi
Jinja2
joblib
jsonlines
jsonpatch
jsonpointer
jsonschema
jupyter_client
jupyter_core
kiwisolver
langchain
langchain-community
langchain-core
langchain-experimental
langchain-google-genai
langchain-openai
langgraph
langsmith
loguru
lxml
markdown-it-py
MarkupSafe
marshmallow
matplotlib
matplotlib-inline
mdurl
mmh3
mpmath
msgpack
multidict
mypy-extensions
nbformat
nest-asyncio
networkx
nomic
numpy
oauthlib
onnxruntime
openai
orjson
overrides
packaging
pandas
parso
peewee
pillow
platformdirs
posthog
prompt_toolkit
protobuf==4.25.3  # version compatible with Python 3.10
psutil
pure_eval
pyarrow
pyasn1
pyasn1_modules
pydantic
pydantic-settings
Pygments
PyJWT
pyparsing
PyPika
python-dateutil
python-dotenv
pytz
PyYAML
pyzmq
referencing
regex
requests
requests-oauthlib
requests-toolbelt
rich
rpds-py
rsa
safetensors
scikit-learn
scipy
sentence-transformers
shellingham
six
sniffio
soupsieve
SQLAlchemy
starlette
sympy
tenacity
threadpoolctl
tiktoken
tokenizers
tomli
torch
tornado
tqdm
traitlets
transformers
typer
typing_extensions
tzdata
uritemplate
urllib3
uvicorn
watchfiles
wcwidth
webencodings
websocket-client
websockets
wikipedia
wrapt
yarl
yfinance
youtube-search
zipp
zstandard
streamlit
-e .
